refactor(shifu): isolate promotion validation and filesystem
