+++ crates/shifu/src/promote.rs
@@ -39,9 +39,14 @@ use crate::artifact_catalog::{
 use crate::native_update::{artifact_sha256, local_artifact_identity_valid};
 use crate::{envfile, native_update, util};
 
+#[path = "promote_catalog.rs"]
+mod promote_catalog;
 #[path = "promote_desktop.rs"]
 mod promote_desktop;
+#[path = "promote_desktop_fs.rs"]
+mod promote_desktop_fs;
 use promote_desktop::*;
+use promote_desktop_fs::DesktopCommit;
 
 struct BuildEntry {
     slot: PathBuf,
@@ -322,94 +327,15 @@ fn current_payload_valid(entry: &BuildEntry) -> bool {
 /// `builds --verify-current` call `build_valid` and re-check the selected
 /// payload in full before making any safety claim or state change.
 fn build_recorded_valid(entry: &BuildEntry) -> bool {
-    !entry.sha.is_empty()
-        && entry.sha != "unknown"
-        && !entry.sha.ends_with("-dirty")
-        && matches!(entry.kind.as_str(), "app" | "installer" | "appimage")
-        && entry.digest.len() == 64
-        && entry.digest.bytes().all(|byte| byte.is_ascii_hexdigit())
-        && !entry.cli_archive.is_empty()
-        && valid_root(&entry.cli_archive_digest)
-        && entry.slot.join(&entry.cli_archive).is_file()
-        && !entry.upgrade_manifest.is_empty()
-        && valid_root(&entry.upgrade_manifest_digest)
-        && entry.slot.join(&entry.upgrade_manifest).is_file()
-        && !entry.product_version.is_empty()
-        && valid_root(&entry.release_cut_root)
-        && valid_root(&entry.platform_slice_root)
-        && entry.mainline_ref == product_mainline_ref()
-        && entry.integrated
-        && entry.qualified
-        && entry.sha == entry.mainline_sha
+    promote_catalog::recorded_build_valid(entry)
 }
 
 fn build_previewable(entry: &BuildEntry) -> bool {
-    !entry.sha.is_empty()
-        && entry.sha != "unknown"
-        && !entry.sha.ends_with("-dirty")
-        && matches!(entry.kind.as_str(), "app" | "installer" | "appimage")
-        && local_release_evidence_valid(entry)
-        && !entry.cli_archive.is_empty()
-        && valid_root(&entry.cli_archive_digest)
-        && entry.slot.join(&entry.cli_archive).is_file()
-        && !entry.upgrade_manifest.is_empty()
-        && valid_root(&entry.upgrade_manifest_digest)
-        && entry.slot.join(&entry.upgrade_manifest).is_file()
-        && !entry.product_version.is_empty()
-        && valid_root(&entry.release_cut_root)
-        && valid_root(&entry.platform_slice_root)
-        && entry.mainline_ref == product_mainline_ref()
-        && product_manifests_valid(entry)
-}
-
-fn local_release_evidence_valid(entry: &BuildEntry) -> bool {
-    let artifact = entry.slot.join(&entry.artifact);
-    let archive = entry.slot.join(&entry.cli_archive);
-    let manifest = entry.slot.join(&entry.upgrade_manifest);
-    if entry.digest.len() != 64
-        || !entry.digest.bytes().all(|byte| byte.is_ascii_hexdigit())
-        || artifact_sha256(&artifact).ok().as_deref() != Some(entry.digest.as_str())
-        || bootstrap::sha256_file(&archive)
-            .ok()
-            .map(|digest| format!("sha256:{digest}"))
-            .as_deref()
-            != Some(entry.cli_archive_digest.as_str())
-        || bootstrap::sha256_file(&manifest)
-            .ok()
-            .map(|digest| format!("sha256:{digest}"))
-            .as_deref()
-            != Some(entry.upgrade_manifest_digest.as_str())
-    {
-        return false;
-    }
-    let Ok(text) = fs::read_to_string(manifest) else {
-        return false;
-    };
-    let Ok(document) = json::parse(&text) else {
-        return false;
-    };
-    let local = document.get("localArtifact");
-    document.str_of("schema") == "kungfu.product-upgrade.manifest/v1"
-        && document.str_of("releaseCutRoot") == entry.release_cut_root
-        && document.str_of("platformSliceRoot") == entry.platform_slice_root
-        && local
-            .map(|value| local_artifact_identity_valid(&artifact, &entry.digest, value))
-            .unwrap_or(false)
+    promote_catalog::previewable_build(entry)
 }
 
 fn valid_root(value: &str) -> bool {
-    value.len() == 71
-        && value.starts_with("sha256:")
-        && value[7..]
-            .bytes()
-            .all(|byte| byte.is_ascii_digit() || (b'a'..=b'f').contains(&byte))
-}
-
-fn product_manifests_valid(entry: &BuildEntry) -> bool {
-    if entry.kind != "app" {
-        return true;
-    }
-    product_app_manifests_valid(&entry.slot.join(&entry.artifact), &entry.sha).unwrap_or(false)
+    promote_catalog::content_root_valid(value)
 }
 
 fn rollback_entry_valid(registry: &Path, build_id: &str, sha: &str) -> bool {
